// IM_FRONTENDA/src/app/models/VistaCliente.ts
export class VistaCliente {
    clienteId: number;
    codigocliente: string | null;
    razonsocial: string | null;
    documentoId: number | null;
    tipoDocumento: string | null;
    numerodocumento: string | null;
    direccion: string | null;
    telefono: string | null;
    email: string | null;

constructor() {

    this.clienteId=0;
    this.codigocliente="";
    this.razonsocial="";
    this.documentoId=0;
    this.tipoDocumento="";
    this.numerodocumento="";
    this.direccion="";
    this.telefono="";
    this.email="";


}

}